fix: add order by to make values be picked up
ST_Value returns a table with one col and multiple rows. A row for each raster tile that ST_Intersects  returns true for. Previously, all except one value at tile borders were null. The valid value was at the bottom of the table returned by ST_Value. Only the topmost point is picked up by the application.
The order by now makes sure that valid value is at the top of the table in order to be picked up.